Rename "typings" to "types" in solid-components
