+++ packages/solid-components/src/base/form/input-field.tsx
@@ -1,5 +1,5 @@
 import { assignProps, Component } from 'solid-js';
-import { DefaultComponents } from '../../../typings/component-creater';
+import { DefaultComponents } from '../../../types/component-creater';
 
 type DefaultPropsMap = {
   wrapper: JSX.HTMLAttributes<HTMLDivElement>;
