+++ packages/solid-components/src/cirrus/common/basic-input-field.tsx
@@ -1,8 +1,8 @@
 import clsx, { Clsx } from 'clsx';
 import { assignProps, Component } from 'solid-js';
 import { sizeSuffixMap } from '../../util/cirrus-utils';
-import { SizedFormItem } from '../../../typings/cirrus-form';
-import { Cirrus } from '../../../typings/cirrus-style';
+import { SizedFormItem } from '../../../types/cirrus-form';
+import { Cirrus } from '../../../types/cirrus-style';
 
 const cn: Clsx<Cirrus> = clsx;
