+++ packages/solid-components/src/cirrus/common/header-dropdown.tsx
@@ -2,8 +2,8 @@ import { css } from 'styled-jsx/css';
 import { createSignal, For, assignProps } from 'solid-js';
 import clsx, { Clsx } from 'clsx';
 import { Portal } from 'solid-js/web';
-import { Cirrus } from '../../../typings/cirrus-style';
-import { ComponentCreater } from '../../../typings/component-creater';
+import { Cirrus } from '../../../types/cirrus-style';
+import { ComponentCreater } from '../../../types/component-creater';
 import { buttonize } from '../../util/component-utils';
 
 type LinkButtonClass = 'LinkButtonClassName';
