+++ packages/solid-components/src/cirrus/domain/login-basic-bottom.tsx
@@ -1,7 +1,7 @@
 import clsx, { Clsx } from 'clsx';
 import { assignProps, Component } from 'solid-js';
-import { Cirrus } from '../../../typings/cirrus-style';
-import { DefaultComponents } from '../../../typings/component-creater';
+import { Cirrus } from '../../../types/cirrus-style';
+import { DefaultComponents } from '../../../types/component-creater';
 
 const cn: Clsx<Cirrus> = clsx;
