+++ packages/solid-components/src/util/input-field-utils.ts
@@ -1,5 +1,5 @@
-import { NativeHandlerOf } from '@components/common/typings/component-utils';
 import { batch } from 'solid-js';
+import { NativeHandlerOf } from '../../types/component-utils';
 
 export const inputRegexSource = {
   // email: RegExp('^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$')
