+++ packages/solid-components/types/firebase-cdn.d.ts
