+++ packages/solid-components/types/firebase-ui-cdn.d.ts
@@ -1,0 +1,5 @@
+import ui from 'firebaseui';
+
+declare global {
+  const firebaseui: typeof ui;
+}
