--- packages/solid-components/typings/cirrus-style.d.ts
@@ -1,989 +1,0 @@
-export type Cirrus =
-  | 'action-bar'
-  | 'active'
-  | 'alternate'
-  | 'animated'
-  | 'avatar'
-  | 'avatar--large'
-  | 'avatar--small'
-  | 'avatar--xlarge'
-  | 'avatar--xsmall'
-  | 'bg-blue-100'
-  | 'bg-blue-200'
-  | 'bg-blue-300'
-  | 'bg-blue-400'
-  | 'bg-blue-500'
-  | 'bg-blue-600'
-  | 'bg-blue-700'
-  | 'bg-blue-800'
-  | 'bg-blue-900'
-  | 'bg-danger'
-  | 'bg-dark'
-  | 'bg-gray-000'
-  | 'bg-gray-100'
-  | 'bg-gray-200'
-  | 'bg-gray-300'
-  | 'bg-gray-400'
-  | 'bg-gray-500'
-  | 'bg-gray-600'
-  | 'bg-gray-700'
-  | 'bg-gray-800'
-  | 'bg-gray-900'
-  | 'bg-green-100'
-  | 'bg-green-200'
-  | 'bg-green-300'
-  | 'bg-green-400'
-  | 'bg-green-500'
-  | 'bg-green-600'
-  | 'bg-green-700'
-  | 'bg-green-800'
-  | 'bg-green-900'
-  | 'bg-indigo-100'
-  | 'bg-indigo-200'
-  | 'bg-indigo-300'
-  | 'bg-indigo-400'
-  | 'bg-indigo-500'
-  | 'bg-indigo-600'
-  | 'bg-indigo-700'
-  | 'bg-indigo-800'
-  | 'bg-indigo-900'
-  | 'bg-info'
-  | 'bg-light'
-  | 'bg-link'
-  | 'bg-link-dark'
-  | 'bg-orange-100'
-  | 'bg-orange-200'
-  | 'bg-orange-300'
-  | 'bg-orange-400'
-  | 'bg-orange-500'
-  | 'bg-orange-600'
-  | 'bg-orange-700'
-  | 'bg-orange-800'
-  | 'bg-orange-900'
-  | 'bg-pink-100'
-  | 'bg-pink-200'
-  | 'bg-pink-300'
-  | 'bg-pink-400'
-  | 'bg-pink-500'
-  | 'bg-pink-600'
-  | 'bg-pink-700'
-  | 'bg-pink-800'
-  | 'bg-pink-900'
-  | 'bg-primary'
-  | 'bg-purple-100'
-  | 'bg-purple-200'
-  | 'bg-purple-300'
-  | 'bg-purple-400'
-  | 'bg-purple-500'
-  | 'bg-purple-600'
-  | 'bg-purple-700'
-  | 'bg-purple-800'
-  | 'bg-purple-900'
-  | 'bg-red-100'
-  | 'bg-red-200'
-  | 'bg-red-300'
-  | 'bg-red-400'
-  | 'bg-red-500'
-  | 'bg-red-600'
-  | 'bg-red-700'
-  | 'bg-red-800'
-  | 'bg-red-900'
-  | 'bg-success'
-  | 'bg-teal-100'
-  | 'bg-teal-200'
-  | 'bg-teal-300'
-  | 'bg-teal-400'
-  | 'bg-teal-500'
-  | 'bg-teal-600'
-  | 'bg-teal-700'
-  | 'bg-teal-800'
-  | 'bg-teal-900'
-  | 'bg-warning'
-  | 'bg-yellow-100'
-  | 'bg-yellow-200'
-  | 'bg-yellow-300'
-  | 'bg-yellow-400'
-  | 'bg-yellow-500'
-  | 'bg-yellow-600'
-  | 'bg-yellow-700'
-  | 'bg-yellow-800'
-  | 'bg-yellow-900'
-  | 'bordered'
-  | 'borderless'
-  | 'bounce'
-  | 'bounceIn'
-  | 'btn'
-  | 'btn--circle'
-  | 'btn--pilled'
-  | 'btn-accent'
-  | 'btn-animated'
-  | 'btn-black'
-  | 'btn-close'
-  | 'btn-danger'
-  | 'btn-dark'
-  | 'btn-dropdown'
-  | 'btn-error'
-  | 'btn-group'
-  | 'btn-group-fill'
-  | 'btn-info'
-  | 'btn-large'
-  | 'btn-light'
-  | 'btn-link'
-  | 'btn-primary'
-  | 'btn-small'
-  | 'btn-success'
-  | 'btn-transparent'
-  | 'btn-warning'
-  | 'btn-xlarge'
-  | 'btn-xsmall'
-  | 'capitalize'
-  | 'card'
-  | 'card-body'
-  | 'card-container'
-  | 'card-footer'
-  | 'card-head'
-  | 'card-head-title'
-  | 'card-image'
-  | 'card-tile'
-  | 'caret'
-  | 'col'
-  | 'col-1'
-  | 'col-10'
-  | 'col-11'
-  | 'col-12'
-  | 'col-2'
-  | 'col-3'
-  | 'col-4'
-  | 'col-5'
-  | 'col-6'
-  | 'col-7'
-  | 'col-8'
-  | 'col-9'
-  | 'col-lg-1'
-  | 'col-lg-10'
-  | 'col-lg-11'
-  | 'col-lg-12'
-  | 'col-lg-2'
-  | 'col-lg-3'
-  | 'col-lg-4'
-  | 'col-lg-5'
-  | 'col-lg-6'
-  | 'col-lg-7'
-  | 'col-lg-8'
-  | 'col-lg-9'
-  | 'col-md-1'
-  | 'col-md-10'
-  | 'col-md-11'
-  | 'col-md-12'
-  | 'col-md-2'
-  | 'col-md-3'
-  | 'col-md-4'
-  | 'col-md-5'
-  | 'col-md-6'
-  | 'col-md-7'
-  | 'col-md-8'
-  | 'col-md-9'
-  | 'col-sm-1'
-  | 'col-sm-10'
-  | 'col-sm-11'
-  | 'col-sm-12'
-  | 'col-sm-2'
-  | 'col-sm-3'
-  | 'col-sm-4'
-  | 'col-sm-5'
-  | 'col-sm-6'
-  | 'col-sm-7'
-  | 'col-sm-8'
-  | 'col-sm-9'
-  | 'col-xs-1'
-  | 'col-xs-10'
-  | 'col-xs-11'
-  | 'col-xs-12'
-  | 'col-xs-2'
-  | 'col-xs-3'
-  | 'col-xs-4'
-  | 'col-xs-5'
-  | 'col-xs-6'
-  | 'col-xs-7'
-  | 'col-xs-8'
-  | 'col-xs-9'
-  | 'container'
-  | 'content'
-  | 'dark'
-  | 'delay'
-  | 'disabled'
-  | 'divided'
-  | 'divider'
-  | 'divider--v'
-  | 'dropdown-animated'
-  | 'dropdown-dark'
-  | 'dropdown-menu'
-  | 'dropdown-menu-divider'
-  | 'dropdown-right'
-  | 'dropdown-shown'
-  | 'ellipses'
-  | 'fa-wrapper'
-  | 'fadeIn'
-  | 'faded'
-  | 'fig'
-  | 'fig-caption'
-  | 'fill-height'
-  | 'fixed-head'
-  | 'font-alt'
-  | 'font-bold'
-  | 'font-italic'
-  | 'font-light'
-  | 'font-normal'
-  | 'font-semibold'
-  | 'font-thin'
-  | 'footer'
-  | 'footer--fixed'
-  | 'footer__list-item'
-  | 'footer__list-title'
-  | 'footer__title'
-  | 'form-ext-checkbox'
-  | 'form-ext-control'
-  | 'form-ext-input'
-  | 'form-ext-input--danger'
-  | 'form-ext-input--dark'
-  | 'form-ext-input--gray'
-  | 'form-ext-input--info'
-  | 'form-ext-input--link'
-  | 'form-ext-input--success'
-  | 'form-ext-input--warning'
-  | 'form-ext-label'
-  | 'form-ext-radio'
-  | 'form-ext-toggle'
-  | 'form-ext-toggle--danger'
-  | 'form-ext-toggle--dark'
-  | 'form-ext-toggle--gray'
-  | 'form-ext-toggle--info'
-  | 'form-ext-toggle--link'
-  | 'form-ext-toggle--success'
-  | 'form-ext-toggle--warning'
-  | 'form-ext-toggle__label'
-  | 'form-ext-toggle__toggler'
-  | 'form-group'
-  | 'form-group-btn'
-  | 'form-group-input'
-  | 'form-group-label'
-  | 'frame'
-  | 'frame__body'
-  | 'frame__footer'
-  | 'frame__header'
-  | 'frame__nav'
-  | 'frame__subtitle'
-  | 'frame__title'
-  | 'fullscreen'
-  | 'grid'
-  | 'grid-c-1'
-  | 'grid-c-10'
-  | 'grid-c-11'
-  | 'grid-c-12'
-  | 'grid-c-2'
-  | 'grid-c-3'
-  | 'grid-c-4'
-  | 'grid-c-5'
-  | 'grid-c-6'
-  | 'grid-c-7'
-  | 'grid-c-8'
-  | 'grid-c-9'
-  | 'grid-ce-1'
-  | 'grid-ce-10'
-  | 'grid-ce-11'
-  | 'grid-ce-12'
-  | 'grid-ce-2'
-  | 'grid-ce-3'
-  | 'grid-ce-4'
-  | 'grid-ce-5'
-  | 'grid-ce-6'
-  | 'grid-ce-7'
-  | 'grid-ce-8'
-  | 'grid-ce-9'
-  | 'grid-ce-auto'
-  | 'grid-ce-end'
-  | 'grid-cols-1'
-  | 'grid-cols-10'
-  | 'grid-cols-11'
-  | 'grid-cols-12'
-  | 'grid-cols-2'
-  | 'grid-cols-3'
-  | 'grid-cols-4'
-  | 'grid-cols-5'
-  | 'grid-cols-6'
-  | 'grid-cols-7'
-  | 'grid-cols-8'
-  | 'grid-cols-9'
-  | 'grid-cs-1'
-  | 'grid-cs-10'
-  | 'grid-cs-11'
-  | 'grid-cs-12'
-  | 'grid-cs-2'
-  | 'grid-cs-3'
-  | 'grid-cs-4'
-  | 'grid-cs-5'
-  | 'grid-cs-6'
-  | 'grid-cs-7'
-  | 'grid-cs-8'
-  | 'grid-cs-9'
-  | 'grid-gap-1'
-  | 'grid-gap-10'
-  | 'grid-gap-11'
-  | 'grid-gap-12'
-  | 'grid-gap-2'
-  | 'grid-gap-3'
-  | 'grid-gap-4'
-  | 'grid-gap-5'
-  | 'grid-gap-6'
-  | 'grid-gap-7'
-  | 'grid-gap-8'
-  | 'grid-gap-9'
-  | 'grid-r-1'
-  | 'grid-r-10'
-  | 'grid-r-11'
-  | 'grid-r-12'
-  | 'grid-r-2'
-  | 'grid-r-3'
-  | 'grid-r-4'
-  | 'grid-r-5'
-  | 'grid-r-6'
-  | 'grid-r-7'
-  | 'grid-r-8'
-  | 'grid-r-9'
-  | 'grid-re-1'
-  | 'grid-re-10'
-  | 'grid-re-11'
-  | 'grid-re-12'
-  | 'grid-re-2'
-  | 'grid-re-3'
-  | 'grid-re-4'
-  | 'grid-re-5'
-  | 'grid-re-6'
-  | 'grid-re-7'
-  | 'grid-re-8'
-  | 'grid-re-9'
-  | 'grid-re-auto'
-  | 'grid-re-end'
-  | 'grid-rs-1'
-  | 'grid-rs-10'
-  | 'grid-rs-11'
-  | 'grid-rs-12'
-  | 'grid-rs-2'
-  | 'grid-rs-3'
-  | 'grid-rs-4'
-  | 'grid-rs-5'
-  | 'grid-rs-6'
-  | 'grid-rs-7'
-  | 'grid-rs-8'
-  | 'grid-rs-9'
-  | 'group-tags'
-  | 'h-10'
-  | 'h-100'
-  | 'h-20'
-  | 'h-30'
-  | 'h-40'
-  | 'h-50'
-  | 'h-60'
-  | 'h-70'
-  | 'h-80'
-  | 'h-90'
-  | 'h-auto'
-  | 'has-sub'
-  | 'header'
-  | 'header-animated'
-  | 'header-brand'
-  | 'header-clear'
-  | 'header-dark'
-  | 'header-fixed'
-  | 'header-nav'
-  | 'headline-1'
-  | 'headline-2'
-  | 'headline-3'
-  | 'headline-4'
-  | 'hero'
-  | 'hero-body'
-  | 'hero-img'
-  | 'hide-text'
-  | 'hover-grow'
-  | 'hovered'
-  | 'icon'
-  | 'icon-left'
-  | 'icon-right'
-  | 'ignore-screen'
-  | 'img-contain'
-  | 'img-cover'
-  | 'img-stretch'
-  | 'infinite'
-  | 'info'
-  | 'inline'
-  | 'input-contains-icon'
-  | 'input-contains-icon-left'
-  | 'input-contains-icon-right'
-  | 'input-control'
-  | 'input-control--pilled'
-  | 'input-error'
-  | 'input-focused'
-  | 'input-large'
-  | 'input-small'
-  | 'input-success'
-  | 'input-xlarge'
-  | 'input-xsmall'
-  | 'label'
-  | 'label-large'
-  | 'label-small'
-  | 'label-xlarge'
-  | 'label-xsmall'
-  | 'large'
-  | 'lead'
-  | 'level'
-  | 'level-content'
-  | 'level-item'
-  | 'level-left'
-  | 'level-right'
-  | 'list-dropdown'
-  | 'loading'
-  | 'loading-left'
-  | 'loading-right'
-  | 'loading-white'
-  | 'lowercase'
-  | 'm-0'
-  | 'm-1'
-  | 'm-10'
-  | 'm-12'
-  | 'm-16'
-  | 'm-2'
-  | 'm-20'
-  | 'm-24'
-  | 'm-3'
-  | 'm-32'
-  | 'm-4'
-  | 'm-5'
-  | 'm-6'
-  | 'm-8'
-  | 'mark'
-  | 'mb-0'
-  | 'mb-1'
-  | 'mb-10'
-  | 'mb-12'
-  | 'mb-16'
-  | 'mb-2'
-  | 'mb-20'
-  | 'mb-24'
-  | 'mb-3'
-  | 'mb-32'
-  | 'mb-4'
-  | 'mb-5'
-  | 'mb-6'
-  | 'mb-8'
-  | 'mb-auto'
-  | 'media-stretch'
-  | 'menu'
-  | 'menu-addon'
-  | 'menu-item'
-  | 'menu-title'
-  | 'ml-0'
-  | 'ml-1'
-  | 'ml-10'
-  | 'ml-12'
-  | 'ml-16'
-  | 'ml-2'
-  | 'ml-20'
-  | 'ml-24'
-  | 'ml-3'
-  | 'ml-32'
-  | 'ml-4'
-  | 'ml-5'
-  | 'ml-6'
-  | 'ml-8'
-  | 'ml-auto'
-  | 'mobile-title'
-  | 'modal'
-  | 'modal-animated--dropdown'
-  | 'modal-animated--zoom-in'
-  | 'modal-animated--zoom-out'
-  | 'modal-body'
-  | 'modal-container'
-  | 'modal-content'
-  | 'modal-footer'
-  | 'modal-header'
-  | 'modal-large'
-  | 'modal-overlay'
-  | 'modal-small'
-  | 'modal-title'
-  | 'mr-0'
-  | 'mr-1'
-  | 'mr-10'
-  | 'mr-12'
-  | 'mr-16'
-  | 'mr-2'
-  | 'mr-20'
-  | 'mr-24'
-  | 'mr-3'
-  | 'mr-32'
-  | 'mr-4'
-  | 'mr-5'
-  | 'mr-6'
-  | 'mr-8'
-  | 'mr-auto'
-  | 'mt-0'
-  | 'mt-1'
-  | 'mt-10'
-  | 'mt-12'
-  | 'mt-16'
-  | 'mt-2'
-  | 'mt-20'
-  | 'mt-24'
-  | 'mt-3'
-  | 'mt-32'
-  | 'mt-4'
-  | 'mt-5'
-  | 'mt-6'
-  | 'mt-8'
-  | 'mt-auto'
-  | 'mx-0'
-  | 'mx-1'
-  | 'mx-10'
-  | 'mx-12'
-  | 'mx-16'
-  | 'mx-2'
-  | 'mx-20'
-  | 'mx-24'
-  | 'mx-3'
-  | 'mx-32'
-  | 'mx-4'
-  | 'mx-5'
-  | 'mx-6'
-  | 'mx-8'
-  | 'mx-auto'
-  | 'my-0'
-  | 'my-1'
-  | 'my-10'
-  | 'my-12'
-  | 'my-16'
-  | 'my-2'
-  | 'my-20'
-  | 'my-24'
-  | 'my-3'
-  | 'my-32'
-  | 'my-4'
-  | 'my-5'
-  | 'my-6'
-  | 'my-8'
-  | 'my-auto'
-  | 'nav-btn'
-  | 'nav-center'
-  | 'nav-dropdown-link'
-  | 'nav-item'
-  | 'nav-left'
-  | 'nav-menu'
-  | 'nav-overflow-x'
-  | 'nav-right'
-  | 'no-bullets'
-  | 'no-space'
-  | 'offset-1'
-  | 'offset-10'
-  | 'offset-11'
-  | 'offset-12'
-  | 'offset-2'
-  | 'offset-3'
-  | 'offset-4'
-  | 'offset-5'
-  | 'offset-6'
-  | 'offset-7'
-  | 'offset-8'
-  | 'offset-9'
-  | 'offset-center'
-  | 'offset-left'
-  | 'offset-right'
-  | 'outline'
-  | 'p-0'
-  | 'p-1'
-  | 'p-10'
-  | 'p-12'
-  | 'p-16'
-  | 'p-2'
-  | 'p-20'
-  | 'p-24'
-  | 'p-3'
-  | 'p-32'
-  | 'p-4'
-  | 'p-5'
-  | 'p-6'
-  | 'p-8'
-  | 'pad-left'
-  | 'pad-right'
-  | 'padded'
-  | 'pagination'
-  | 'pagination-bordered'
-  | 'pagination-item'
-  | 'pagination-item-subtitle'
-  | 'pagination-next'
-  | 'pagination-prev'
-  | 'parallax-img'
-  | 'paused'
-  | 'pb-0'
-  | 'pb-1'
-  | 'pb-10'
-  | 'pb-12'
-  | 'pb-16'
-  | 'pb-2'
-  | 'pb-20'
-  | 'pb-24'
-  | 'pb-3'
-  | 'pb-32'
-  | 'pb-4'
-  | 'pb-5'
-  | 'pb-6'
-  | 'pb-8'
-  | 'pl-0'
-  | 'pl-1'
-  | 'pl-10'
-  | 'pl-12'
-  | 'pl-16'
-  | 'pl-2'
-  | 'pl-20'
-  | 'pl-24'
-  | 'pl-3'
-  | 'pl-32'
-  | 'pl-4'
-  | 'pl-5'
-  | 'pl-6'
-  | 'pl-8'
-  | 'placeholder'
-  | 'placeholder-commands'
-  | 'placeholder-icon'
-  | 'placeholder-subtitle'
-  | 'placeholder-title'
-  | 'pr-0'
-  | 'pr-1'
-  | 'pr-10'
-  | 'pr-12'
-  | 'pr-16'
-  | 'pr-2'
-  | 'pr-20'
-  | 'pr-24'
-  | 'pr-3'
-  | 'pr-32'
-  | 'pr-4'
-  | 'pr-5'
-  | 'pr-6'
-  | 'pr-8'
-  | 'pt-0'
-  | 'pt-1'
-  | 'pt-10'
-  | 'pt-12'
-  | 'pt-16'
-  | 'pt-2'
-  | 'pt-20'
-  | 'pt-24'
-  | 'pt-3'
-  | 'pt-32'
-  | 'pt-4'
-  | 'pt-5'
-  | 'pt-6'
-  | 'pt-8'
-  | 'pulse'
-  | 'px-0'
-  | 'px-1'
-  | 'px-10'
-  | 'px-12'
-  | 'px-16'
-  | 'px-2'
-  | 'px-20'
-  | 'px-24'
-  | 'px-3'
-  | 'px-32'
-  | 'px-4'
-  | 'px-5'
-  | 'px-6'
-  | 'px-8'
-  | 'py-0'
-  | 'py-1'
-  | 'py-10'
-  | 'py-12'
-  | 'py-16'
-  | 'py-2'
-  | 'py-20'
-  | 'py-24'
-  | 'py-3'
-  | 'py-32'
-  | 'py-4'
-  | 'py-5'
-  | 'py-6'
-  | 'py-8'
-  | 'r'
-  | 'rat-1-1'
-  | 'rat-4-3'
-  | 'required'
-  | 'right'
-  | 'row'
-  | 'row--no-wrap'
-  | 'rtl'
-  | 'search'
-  | 'select'
-  | 'selected'
-  | 'short'
-  | 'shown'
-  | 'slide-up'
-  | 'small'
-  | 'space'
-  | 'striped'
-  | 'subtitle'
-  | 'tab-container'
-  | 'tab-item-content'
-  | 'table'
-  | 'tabs-center'
-  | 'tabs-classic'
-  | 'tabs-depth'
-  | 'tabs-fill'
-  | 'tabs-large'
-  | 'tabs-right'
-  | 'tabs-small'
-  | 'tabs-xlarge'
-  | 'tabs-xsmall'
-  | 'tag'
-  | 'tag--black'
-  | 'tag--danger'
-  | 'tag--dark'
-  | 'tag--gray'
-  | 'tag--info'
-  | 'tag--large'
-  | 'tag--link'
-  | 'tag--primary'
-  | 'tag--rounded'
-  | 'tag--success'
-  | 'tag--warning'
-  | 'tag--white'
-  | 'tag--xlarge'
-  | 'tag-container'
-  | 'tag-container--centered'
-  | 'tag-container--grouped'
-  | 'tag-container--right'
-  | 'tag__close-btn'
-  | 'tag__delete'
-  | 'text--large'
-  | 'text-blue-100'
-  | 'text-blue-200'
-  | 'text-blue-300'
-  | 'text-blue-400'
-  | 'text-blue-500'
-  | 'text-blue-600'
-  | 'text-blue-700'
-  | 'text-blue-800'
-  | 'text-blue-900'
-  | 'text-danger'
-  | 'text-dark'
-  | 'text-gray-000'
-  | 'text-gray-100'
-  | 'text-gray-200'
-  | 'text-gray-300'
-  | 'text-gray-400'
-  | 'text-gray-500'
-  | 'text-gray-600'
-  | 'text-gray-700'
-  | 'text-gray-800'
-  | 'text-gray-900'
-  | 'text-green-100'
-  | 'text-green-200'
-  | 'text-green-300'
-  | 'text-green-400'
-  | 'text-green-500'
-  | 'text-green-600'
-  | 'text-green-700'
-  | 'text-green-800'
-  | 'text-green-900'
-  | 'text-indigo-100'
-  | 'text-indigo-200'
-  | 'text-indigo-300'
-  | 'text-indigo-400'
-  | 'text-indigo-500'
-  | 'text-indigo-600'
-  | 'text-indigo-700'
-  | 'text-indigo-800'
-  | 'text-indigo-900'
-  | 'text-info'
-  | 'text-light'
-  | 'text-link'
-  | 'text-link-dark'
-  | 'text-orange-100'
-  | 'text-orange-200'
-  | 'text-orange-300'
-  | 'text-orange-400'
-  | 'text-orange-500'
-  | 'text-orange-600'
-  | 'text-orange-700'
-  | 'text-orange-800'
-  | 'text-orange-900'
-  | 'text-pink-100'
-  | 'text-pink-200'
-  | 'text-pink-300'
-  | 'text-pink-400'
-  | 'text-pink-500'
-  | 'text-pink-600'
-  | 'text-pink-700'
-  | 'text-pink-800'
-  | 'text-pink-900'
-  | 'text-primary'
-  | 'text-purple-100'
-  | 'text-purple-200'
-  | 'text-purple-300'
-  | 'text-purple-400'
-  | 'text-purple-500'
-  | 'text-purple-600'
-  | 'text-purple-700'
-  | 'text-purple-800'
-  | 'text-purple-900'
-  | 'text-red-100'
-  | 'text-red-200'
-  | 'text-red-300'
-  | 'text-red-400'
-  | 'text-red-500'
-  | 'text-red-600'
-  | 'text-red-700'
-  | 'text-red-800'
-  | 'text-red-900'
-  | 'text-success'
-  | 'text-teal-100'
-  | 'text-teal-200'
-  | 'text-teal-300'
-  | 'text-teal-400'
-  | 'text-teal-500'
-  | 'text-teal-600'
-  | 'text-teal-700'
-  | 'text-teal-800'
-  | 'text-teal-900'
-  | 'text-warning'
-  | 'text-yellow-100'
-  | 'text-yellow-200'
-  | 'text-yellow-300'
-  | 'text-yellow-400'
-  | 'text-yellow-500'
-  | 'text-yellow-600'
-  | 'text-yellow-700'
-  | 'text-yellow-800'
-  | 'text-yellow-900'
-  | 'tile'
-  | 'tile__buttons'
-  | 'tile__container'
-  | 'tile__icon'
-  | 'tile__subtitle'
-  | 'tile__title'
-  | 'title'
-  | 'title-container'
-  | 'toast'
-  | 'toast--danger'
-  | 'toast--dark'
-  | 'toast--gray'
-  | 'toast--info'
-  | 'toast--link'
-  | 'toast--primary'
-  | 'toast--success'
-  | 'toast--translucent'
-  | 'toast--warning'
-  | 'toast__title'
-  | 'toggle-hover'
-  | 'tooltip'
-  | 'tooltip--bottom'
-  | 'tooltip--bottom-left'
-  | 'tooltip--bottom-right'
-  | 'tooltip--left'
-  | 'tooltip--right'
-  | 'tooltip--top-left'
-  | 'tooltip--top-right'
-  | 'tree'
-  | 'tree-item'
-  | 'tree-item-body'
-  | 'tree-item-header'
-  | 'tree-nav'
-  | 'tree-nav-body'
-  | 'tree-nav-close'
-  | 'tree-nav-container'
-  | 'tree-nav-content'
-  | 'tree-nav-header'
-  | 'u'
-  | 'u-C'
-  | 'u-LR'
-  | 'u-RL'
-  | 'u-block'
-  | 'u-center'
-  | 'u-center-alt'
-  | 'u-circle'
-  | 'u-clear-both'
-  | 'u-clear-left'
-  | 'u-clear-right'
-  | 'u-clearfix'
-  | 'u-disabled'
-  | 'u-flex'
-  | 'u-flex-column'
-  | 'u-flex-column-reverse'
-  | 'u-flex-row'
-  | 'u-flex-row-reverse'
-  | 'u-hide-overflow'
-  | 'u-inline'
-  | 'u-inline-block'
-  | 'u-inline-flex'
-  | 'u-items-baseline'
-  | 'u-items-center'
-  | 'u-items-flex-end'
-  | 'u-items-flex-start'
-  | 'u-items-stretch'
-  | 'u-justify-center'
-  | 'u-justify-flex-end'
-  | 'u-justify-flex-start'
-  | 'u-justify-space-around'
-  | 'u-justify-space-between'
-  | 'u-justify-space-evenly'
-  | 'u-no-outline'
-  | 'u-none'
-  | 'u-none-lg'
-  | 'u-none-md'
-  | 'u-none-md-only'
-  | 'u-none-sm'
-  | 'u-none-sm-only'
-  | 'u-none-xs'
-  | 'u-overflow-auto'
-  | 'u-overlay'
-  | 'u-position-absolute'
-  | 'u-position-fixed'
-  | 'u-position-relative'
-  | 'u-position-static'
-  | 'u-position-sticky'
-  | 'u-pull-left'
-  | 'u-pull-right'
-  | 'u-round'
-  | 'u-shadow'
-  | 'u-table'
-  | 'u-table-cell'
-  | 'u-table-column'
-  | 'u-table-row'
-  | 'u-text-break'
-  | 'u-text-center'
-  | 'u-text-ellipsis'
-  | 'u-text-justify'
-  | 'u-text-left'
-  | 'u-text-right'
-  | 'u-unselectable'
-  | 'underline'
-  | 'uppercase'
-  | 'usquare'
-  | 'utb'
-  | 'utb-C'
-  | 'utb-LR'
-  | 'utb-OLR'
-  | 'utb-ORL'
-  | 'utb-RL'
-  | 'video-fullscreen'
-  | 'w-10'
-  | 'w-100'
-  | 'w-20'
-  | 'w-30'
-  | 'w-40'
-  | 'w-50'
-  | 'w-60'
-  | 'w-70'
-  | 'w-80'
-  | 'w-90'
-  | 'w-auto'
-  | 'white'
-  | 'x-small'
-  | 'xlarge';
